website: Set up React web application using create-react-app
At some point, this directory might have to be merged with the server/
directory if we want to use AWS Cloudformation templates (I need to
check the details of this).  Using Cloudformation templates would
allow us to version-control our infrastructure and also make the lives
of future system administrators easier.