website: Allow multiple instances of website each with dedicated server backends (#51)
See `website/BUILD.md` for how to create all the instances of the web interface (since they're all made together in one CloudFormation stack) and how to create a dedicated server backend to serve the instance of the web interface you're working on (including localhost).

See `server/BUILD.md` for more detailed instructions on how to build and deploy the backend, as well as instructions on how to develop for the server.  Backend will need to be built specifically for the domain that the web interface will be hosted on.  For example, localhost and the dev website will need specially-built backends.  See `server/BUILD.md` for more.

Website code is no longer hardcoded with the API URL.  It is now `REACT_APP_API_URL` (create-react-app requires environment variables passed to dev server or the build command to be prefixed with `REACT_APP`).  The code has also been updated to use the new sample field names from the API.

Squashed commits:

* Allow non-wildcard CORS to work on website

* Update sample search code to new API spec

* Don't hardcode API URL

* Add website build instructions

* Add website samconfig.toml

* website: Ignore SAM junk

* App team not "included" for prod website

They don't even use it.

* website: Add cleanup instructions

* server: Add samconfig.toml

* Remove newline

* Set up template to start up all instances of website s3 buckets

* Import URLs from S3 bucket to allow CORS

* Output origin allowed by CORS

* Reorganise samconfig sections

Each usecase will have be in separate stacks and VPCs.  And within
each usecase, for environments: dev and localhost share a
stack (`-dev`) and VPC but the prod environment has its own stack and
VPC.

The stack names have the following format: `backend-<usecase>-dev`or
`backend-<usecase>-prod`.

Currently, we have usecases for:

- server
- app
- website
- all

We only have dev and localhost environments.

* Improve website build instructions

* Specify UseCase and Environment for server-dev

* Add TODO comment to rename app dev stack

* Rename app to app-dev

* Hardcode "dev" stage

* Document how to build backend

* Clarify split between development and production environments

* Clarify extra use of FuncMetaSetupTestData

* Remove outdated comment

* Set up special usecase for ML classifier

* Add sections for debugging and links to help

* Fix typo

* Remove outdated comment

* Use mappings to get Origin for CORS config

Much nicer now.

* Always confirm changesets

* Revert "Use mappings to get Origin for CORS config"

This reverts commit 2da515ccfcf54bf984ef4f2ae890dbca07650dfd.

It didn't work for some reason.

The cloudformation template language sucks.  Just give me Lisp.

* We won't use ACM certificates anymore

It needs a custom domain which our AWS credits can't be used for.

* Remove template to create S3 bucket for subtemplates

Forgot to remove this.  Thank goodness for SAM automatically uploading
artifacts to S3 to these ridiculous restrictions for the templates.

* Fix typo