server: Integrate Cognito auth into backend + misc stuff (#82)
* server: Fix ID generator

Now the ID length and ID generator won't be out of sync.

* server: Integrate backend with Cognito auth

** server: Enable Access-Control-Allow-Credentials

** server: Create Cognito user pool in template

** server: Fix UserPool

The docs say the `Priority` key is optional, but the default is an
invalid value.  How stupid.

** server: Configure email verification with code

** server: Reinstate public internet access for VPC through NAT

*** Revert "server: Remove all public internet access within VPC"

This reverts commit a30d4fc367ff47c6bf56ab82d07cbf2bffd7a84d.

*** server: Update SAM build config to create public subnets

*** server: Use default network ACL for public subnets

Don't need to lock things down there, really.

*** server: Set up NATs

** server: Remove useless comments

** server: Add Cognito dependency to JS layer

** server: Move common config for serverless functions into SAM Globals section

** server: Set up userpool groups and triggers outside of SAM template

Because of circular dependencies.

*** server: Give UserPool permission to execute triggers

*** server: Fix UpdateUserPool Cognito API call

Turns out this API action fully replaces the Cognito userpool config
so we first need to retrieve the config with DescribeUserPool.

We also need to ensure the temporary password validity period is set
in only one place since it can be set in multiple places for some
crazy reason.

** server: Add postconfirm lambda

*** server: Add more concurrency to postconfirm cognito trigger

Updating the DB and the Cognito API call don't depend on each other,
so we can make the network calls concurrently and then handle the
errors after all the promises have "settled".

** server: Add Cognito users to DB tables

User ID is email.

** server: Add action to register researcher accounts

*** server: Fix map-to-object in researcher-register action

JS is a bit quirky since braces in the lambda context means a code
block, so putting an object literal there results in a syntax error
since ":" tokens aren't expected.

** server: Integrate Cognito auth into API actions

*** server: Add helper functions to read cognito token fields

*** server: Use actual ID of caller for children registering

*** server: Check if actually called by parent when registering a child

*** server: Note the potential security vulnerability if requirements change

*** server: Split handlers for child info fetching and modifying

We can choose which function to handle specific events so we can
implement both fetching and modifying in the same source file so it's
easier to keep in sync.

*** server: Ensure only admins and parents of a child can view/modify their info

*** server: Ensure only parents of a child can view their samples

This also adds a `childID` path parameter.

*** server: Ensure only parents of a child can add to their samples

*** server: Refactor authentication code

This is much nicer to look at and to work with.

*** server: Add bash script to test API actions

Only testing auth for now.

*** server: Ensure only admins can make researcher accounts

*** server: Prevent admins from viewing or modifying child personal info

*** server: Add helper lambda to create admin accounts

To reduce the public scope of the API (and therefore improve
security), we will only allow admin accounts to be made on the
backend.

*** server: Test auth for admin in script

*** server: Remove TEMP_PARENT_ID constant and test data lambda

Now that Cognito has been integrated, it doesn't make sense for a test
user in the DB but isn't in Cognito as well.

* server: Delete gender SQL enum when initialising DB

Forgot to do this when I added the gender personal info field.

* server: Return samples from sample-fetching action under `data` property

For consistency with other actions.  This is also specified in the API
document.

* server: Add lambda to clear samples table

Now that we have actual user accounts, we need to be able to clear
samples without removing user IDs from DB (needed for foreign key
relationships).

* server: Add API actions for listing parents, researchers, and admins (separately)

* server: Add API action to list children of a parent